+++ src/components/scrollable_layout.ts
@@ -1,16 +1,5 @@
 import { styled } from "../vendors/stitches.ts";
 
-const defaultScrollbar = {
-  "scrollbarWidth": "initial",
-  "scrollbarColor": "initial",
-  "&::-webkit-scrollbar": { all: "initial" },
-  "&::-webkit-scrollbar-thumb": {
-    all: "initial",
-    background: "#00000088",
-  },
-  "&::-webkit-scrollbar-track": { all: "initial" },
-};
-
 export const Container = styled("div", {
   position: "relative",
 
@@ -35,21 +24,26 @@ export const Container = styled("div", {
   },
 });
 
-export const ScrollableLayout = styled("div", {
+export const OverlayScroller = styled("div", {
   position: "relative",
 
   width: "100%",
   height: "100%",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  flexFlow: "row-reverse wrap",
 
-  overflowY: "auto",
   // For overriding Chrome 118.0.5993.118 UA style sheet
   // :focus-visible { outline: -webkit-focus-ring-color auto 1px; }
   outline: "none",
-  ...defaultScrollbar,
+
+  "& > div[data-overlayscrollbars-viewport]:focus-visible": {
+    // For overriding Chrome 118.0.5993.118 UA style sheet
+    // :focus-visible { outline: -webkit-focus-ring-color auto 1px; }
+    outline: "none",
+  },
+  "& .os-scrollbar-handle": {
+    backdropFilter: "brightness(0.5)",
+    background: "none",
+    border: "#fff8 1px solid",
+  },
 
   variants: {
     fullscreen: {
@@ -58,11 +52,6 @@ export const ScrollableLayout = styled("div", {
         top: 0,
         bottom: 0,
         overflow: "auto",
-      },
-    },
-    ltr: {
-      true: {
-        flexFlow: "row wrap",
       },
     },
     dark: {