+++ src/deps.ts
@@ -11,6 +11,7 @@ export {
   selectAtom,
   splitAtom,
 } from "jotai/utils";
+export { useOverlayScrollbars } from "overlayscrollbars-react";
 export { type Id, toast, ToastContainer } from "react-toastify";
 
 export { Dialog, Tab } from "@headlessui/react";