+++ src/vendors/toastify_css.ts
@@ -1,7 +1,0 @@
-import { insertCss } from "../utils.ts";
-
-insertToastifyCss();
-
-async function insertToastifyCss() {
-  insertCss(await GM.getResourceText("react-toastify-css"));
-}